src: add python wasi example
This commit adds an example of using python from a wasi module.
The goal of this was originally to use numpy from a wasi module but I
ran into an issue as numpy uses a C extension and I was not able to
get that to work. An issue about this is linked in the README.md for
more details.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>